Create stage_upload_folder.sh
This commit adds the bash script `stage_upload.sh` and corresponding instructions in the README file.  It is a standalone script that creates a folder of symbolic links to local files, and is intended to facilitate an upload.  In particular, this approach makes it easier to reorganize and/or rename path and files in whatever manner is desired by the recipient of the transfer.