Add handling for input files without terminal newline
The while loop construct has been rewritten, based on the Stack Overflow post: https://stackoverflow.com/questions/12916352#12919766

This commit was motivated by a problem encountered when the input file did not contain a terminal newline: the while loop would simply skip the last entry in the file.  A test has now been added to allow the loop to proceed if the string in the final line of the file is of non-zero length, even if it is not followed by a newline.